starting objects.ts
